changed uint to unsigned log

+++ hdf5t/dynamic_2d_array.h
@@ -9,7 +9,7 @@ typedef int integer;
 //typedef short sint;
 typedef short sint;
 typedef char cint;
-typedef int tint;
+typedef unsigned long tint;
 typedef float real;
 //typedef double real;
 #define MPI_REALNUM MPI_FLOAT
